+++ projects/Llama/configs/llama_config.py
@@ -48,7 +48,7 @@ cfg = dict(
     eos_token_id=2,
     pad_token_id=0,
     # train
-    pretrained_model_path="/data/home/xiezipeng/hf_models/meta-llama/Llama-2-7b-hf/",
+    pretrained_model_path="Llama-2-7b-hf",
 )
 
 cfg = DictConfig(cfg)
@@ -57,5 +57,5 @@ model = LazyCall(LlamaForCausalLM)(cfg=cfg)
 tokenization = OmegaConf.create()
 tokenization.make_vocab_size_divisible_by = 1
 tokenization.tokenizer = LazyCall(LlamaTokenizer)(
-    pretrained_model_path="/data/home/xiezipeng/hf_models/meta-llama/Llama-2-7b-hf/tokenizer.model"
+    pretrained_model_path="Llama-2-7b-hf/tokenizer.model"
 )
