+++ projects/Llama/configs/llama_sft.py
@@ -18,13 +18,13 @@ from projects.Llama.llama import LlamaForCausalLM
 
 # Hyperparameters
 weight_decay = 0.1
-learning_rate = 2e-5
-max_input_length = 1350
-dataset_path = "/data/home/xiezipeng/libai/projects/Llama/alpaca_data"
-pretrained_model_path = "/data/home/xiezipeng/hf_models/meta-llama/Llama-2-7b-hf/"
+learning_rate = 5e-5
+max_input_length = 512
+dataset_path = "alpaca_data.json"
+pretrained_model_path = "Llama-2-7b-hf"
 
 # graph & optim
-graph["enabled"] = True
+graph["enabled"] = False
 optim.update(
     dict(
         lr=learning_rate,
@@ -47,18 +47,20 @@ dataloader = OmegaConf.create()
 dataloader.train = LazyCall(build_nlp_train_loader)(
     dataset=[
         LazyCall(AlpacaDataset)(
-            path=os.path.join(dataset_path, "train"),
+            path=dataset_path,
             tokenizer=tokenization.tokenizer,
             max_len=max_input_length,
+            partition="train",
         )
     ],
 )
 dataloader.test = [
     LazyCall(build_nlp_test_loader)(
         dataset=LazyCall(AlpacaDataset)(
-            path=os.path.join(dataset_path, "test"),
+            path=dataset_path,
             tokenizer=tokenization.tokenizer,
             max_len=max_input_length,
+            partition="test",
         ),
     ),
 ]
@@ -67,30 +69,30 @@ dataloader.test = [
 train.update(
     dict(
         output_dir="./sft_result",
-        train_micro_batch_size=2,
+        train_micro_batch_size=4,
         test_micro_batch_size=1,
-        train_epoch=5,
+        train_epoch=1,
         train_iter=1,
         log_period=10,
-        warmup_ratio=0,
+        warmup_ratio=2/5,
         num_accumulation_steps=8,
-        rdma_enabled=True,
+        rdma_enabled=False,
         amp=dict(enabled=True),
         activation_checkpoint=dict(enabled=True),
         checkpointer=dict(
-            period=100,
+            period=1000,
             max_to_keep=20,
         ),
         dist=dict(
-            data_parallel_size=2,
-            tensor_parallel_size=1,
-            pipeline_parallel_size=4,
+            data_parallel_size=1,
+            tensor_parallel_size=8,
+            pipeline_parallel_size=1,
             pipeline_num_layers=cfg.hidden_layers,
         ),
         evaluation=dict(
             enabled=True,
             evaluator=LazyCall(PPLEvaluator)(),
-            eval_period=100,
+            eval_period=1000,
             eval_iter=1e5,
         ),
         scheduler=LazyCall(WarmupExponentialLR)(
